moved visualisation files to subdirectory
