+++ app/assets/javascripts/visualisations/cumulative_timeseries.ts
@@ -1,5 +1,5 @@
 import * as d3 from "d3";
-import { RawData, SeriesGraph } from "series_graph";
+import { RawData, SeriesGraph } from "visualisations/series_graph";
 
 export class CTimeseriesGraph extends SeriesGraph {
     protected readonly baseUrl = "/stats/cumulative_timeseries?series_id=";
