+++ app/assets/javascripts/visualisations/series_graph.ts
