+++ app/javascript/packs/visualisations_series.js
@@ -1,4 +1,4 @@
-import { toggleStats, setActiveToggle } from "visualisation_series.js";
+import { toggleStats, setActiveToggle } from "visualisations/visualisation_series.js";
 
 window.dodona.toggleStats = toggleStats;
 window.dodona.setActiveToggle = setActiveToggle;
